requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.4.0
chardet==4.0.0
discord.py==2.2.3
idna==3.3
multidict==6.0.2
python-dotenv==0.20.0
typing_extensions==4.3.0
yarl==1.7.2
numpy==1.23.4
pandas==1.5.1
table2ascii==1.1.0
tabulate==0.9.0
wcwidth==0.2.6
pytz==2022.6